Export TypedString (for now; TODO: move out to a library external to Nevermore)
